// frontend/src/app/api/model/wipStepCreateData.ts
/**
 * My API
 * OpenAPI documentation for My API
 *
 * The version of the OpenAPI document: 59acdd34ea678de65cbe0a25f66df8cefdf5050b
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TemplateStepPosition } from './templateStepPosition';
import { WipStepData } from './wipStepData';
import { WipStepNodeData } from './wipStepNodeData';
import { WipStepPosition } from './wipStepPosition';


export interface WipStepCreateData { 
    stepNumber: string;
    title: string;
    question?: string;
    orderPosition: TemplateStepPosition;
    data: WipStepData;
    incomingConnections: Array<string>;
    outgoingConnections: Array<string>;
    gridPosition: WipStepPosition;
    nodeData: WipStepNodeData;
}